+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/deployment/listener/CarbonUiServer.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.uis.internal;
+package org.wso2.carbon.uis.internal.deployment.listener;
 
 import org.wso2.carbon.uis.api.App;
 import org.wso2.carbon.uis.internal.deployment.AppDeploymentEventListener;
